refactor: 구현클래스 이동
