+++ src/main/kotlin/racingcar/WinnerPrinter.kt
@@ -1,0 +1,14 @@
+package racingcar
+
+import java.io.ByteArrayOutputStream
+import java.io.PrintWriter
+
+class WinnerPrinter(private val names: List<String>, outputStream: ByteArrayOutputStream) {
+    fun print() {
+        printWriter.println(
+            "${names.joinToString(", ")}가 최종 우승했습니다."
+        )
+    }
+
+    private val printWriter = PrintWriter(outputStream, true)
+}
