+++ src/test/kotlin/racingcar/PrinterTest.kt
@@ -3,7 +3,6 @@ package racingcar
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import java.io.ByteArrayOutputStream
-import java.io.PrintWriter
 
 class PrinterTest {
     @Test
@@ -36,15 +35,5 @@ class PrinterTest {
         val outputStream = ByteArrayOutputStream()
         WinnerPrinter(listOf("pobi", "crong"), outputStream).print()
         assertThat(outputStream.toString()).isEqualTo("pobi, crong가 최종 우승했습니다.\n")
-    }
-
-    class WinnerPrinter(private val names: List<String>, outputStream: ByteArrayOutputStream) {
-        fun print() {
-            printWriter.println(
-                "${names.joinToString(", ")}가 최종 우승했습니다."
-            )
-        }
-
-        private val printWriter = PrintWriter(outputStream, true)
     }
 }
